code/haitiwater/requirements.txt
certifi==2018.10.15
chardet==3.0.4
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==2.1.2
django-appconf==1.0.2
django-bootstrap3==11.0.0
django-chartjs==1.3
django-compressor==2.2
django-filter==2.0.0
django-heroku==0.3.1
django-tables2==2.0.2
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
et-xmlfile==1.0.1
gunicorn==19.9.0
heroku==0.1.4
idna==2.7
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.9
psycopg2==2.7.5
python-dateutil==1.5
pytz==2018.6
PyYAML==3.13
rcssmin==1.0.6
requests==2.20.0
rjsmin==1.0.12
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24
virtualenv==16.0.0
whitenoise==4.1
xlrd==1.1.0
xlwt==1.3.0
